time-mcp-server/requirements.txt
fastmcp>=0.1.0
pydantic>=2.5.0
pytz>=2023.3
python-dateutil>=2.8.2
